requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
boto3==1.36.26
python-dotenv==1.0.1
